+++ core/pages/log.php
@@ -15,7 +15,7 @@
  * 
  */
 
-dolibase_include_once('/core/class/page.php');
+dolibase_include_once('/core/class/form_page.php');
 dolibase_include_once('/core/class/logs.php');
 include_once DOL_DOCUMENT_ROOT.'/core/lib/date.lib.php';
 
@@ -23,7 +23,7 @@ include_once DOL_DOCUMENT_ROOT.'/core/lib/date.lib.php';
  * LogPage class
  */
 
-class LogPage extends Page
+class LogPage extends FormPage
 {
 	/**
 	 * Constructor
@@ -45,6 +45,30 @@ class LogPage extends Page
 	}
 
 	/**
+	 * Load default actions
+	 *
+	 */
+	protected function loadDefaultActions()
+	{
+		// Purge search criteria
+		if (GETPOST("button_removefilter_x") || GETPOST("button_removefilter")) // Both test are required to be compatible with all browsers
+		{
+			$_POST = array();
+			//$_GET  = array(); // id & ref should not be removed
+		}
+	}
+
+	/**
+	 * Generate tabs
+	 *
+	 * @param     $noheader     -1 or 0=Add tab header, 1=no tab header.
+	 */
+	protected function generateTabs($noheader = -1)
+	{
+		parent::generateTabs($noheader);
+	}
+
+	/**
 	 * Show banner
 	 *
 	 * @param     $object         object
@@ -59,7 +83,7 @@ class LogPage extends Page
 
 		dol_banner_tab($object, 'ref', $morehtml, 1, 'ref', 'ref', '', '', 0, $morehtmlleft);
 
-		echo '<div class="underbanner clearboth"></div>';
+		//echo '<div class="underbanner clearboth"></div>';
 	}
 
 	/**
@@ -70,8 +94,32 @@ class LogPage extends Page
 	 */
 	public function printLogs($object_id, $object_element = '')
 	{
-		global $langs, $dolibase_config;
+		global $langs, $conf, $dolibase_config;
 
+		// Get parameters
+		$search = array(
+			'ds'   => GETPOSTDATE('ds'), // date start
+			'de'   => GETPOSTDATE('de'), // date end
+			'user' => GETPOST('user')
+		);
+		$sortfield = GETPOST('sortfield', 'alpha');
+		$sortorder = GETPOST('sortorder', 'alpha');
+		$optioncss = GETPOST('optioncss', 'alpha');
+		$limit = GETPOST('limit') ? GETPOST('limit', 'int') : $conf->liste_limit;
+		$page = GETPOST('page', 'int') ? GETPOST('page', 'int') : 0;
+		$offset = $limit * $page;
+		if (empty($sortfield)) $sortfield = "t.datec";
+		if (empty($sortorder)) $sortorder = "DESC";
+
+		// Set list parameters
+		$param = '&id='.$object_id;
+		foreach ($search as $key => $value) {
+			if ($value != '') $param.= '&'.$key.'='.urlencode($value);
+		}
+		if ($optioncss != '') $param.= '&optioncss='.urlencode($optioncss);
+		if ($limit > 0 && $limit != $conf->liste_limit) $param.= '&limit='.urlencode($limit);
+
+		// Fetch logs
 		$log = new Logs();
 		$where = "(module_id = ".$dolibase_config['module']['number'];
 		$where.= " || module_name = '".$dolibase_config['module']['name']."'";
@@ -79,52 +127,92 @@ class LogPage extends Page
 		if (! empty($object_element)) {
 			$where.= " AND object_element = '".$object_element."'";
 		}
+		if ($search['ds'] && $search['de']) $where .= " AND date(t.datec) BETWEEN date('".$log->db->idate($search['ds'])."') AND date('".$log->db->idate($search['de'])."')";
+		if ($search['user'] > 0) $where .= natural_search('t.fk_user', $search['user']);
 
-		// Fetch logs
-		if ($log->fetchAll(0, 0, 't.datec', 'DESC', '', '', $where))
+		$log->fetchAll($limit, $offset, $sortfield, $sortorder, '', '', $where, true);
+
+		// List
+		echo '<form method="POST" action="'.$_SERVER["PHP_SELF"].'?id='.$object_id.'">';
+
+		print_barre_liste('', $page, $_SERVER["PHP_SELF"], $param, $sortfield, $sortorder, '', $log->count, $log->total, '', 0, '', '', $limit);
+
+		echo '<div class="div-table-responsive">';
+		echo '<table class="liste" width="100%">';
+
+		// List header
+		echo '<tr class="liste_titre">';
+		print_liste_field_titre("LogAction", $_SERVER["PHP_SELF"],"t.action","",$param,'align="left"',$sortfield,$sortorder);
+		print_liste_field_titre("LogDate", $_SERVER["PHP_SELF"], "t.datec", "", $param, 'align="center"', $sortfield, $sortorder);
+		print_liste_field_titre("LogUser", $_SERVER["PHP_SELF"], "t.fk_user", "", $param, 'align="left"', $sortfield, $sortorder);
+		print_liste_field_titre('');
+		echo "</tr>\n";
+
+		echo '<tr class="liste_titre">';
+
+		// Action
+		echo '<td align="left" class="liste_titre"></td>';
+
+		// Date
+		echo '<td align="center" class="liste_titre">';
+		echo $this->form->dateInput('ds', $search['ds'], 0);
+		echo $this->form->dateInput('de', $search['de'], 0);
+		echo '</td>';
+
+		// User
+		echo '<td align="left" class="liste_titre">';
+		echo $this->form->select_dolusers($search['user'], 'user', 1, '', 0, '', '', 0, 0, 0, '', 0, '', 'maxwidth300');
+		echo '</td>';
+
+		// Search buttons
+		echo '<td align="right" class="liste_titre">';
+		echo $this->form->showFilterAndCheckAddButtons(0);
+		echo '</td>';
+
+		echo "</tr>\n";
+
+		// Show logs
+		if ($log->count > 0)
 		{
 			$deltadateforserver = getServerTimeZoneInt('now');
 			$deltadateforclient = ((int) $_SESSION['dol_tz'] + (int) $_SESSION['dol_dst']);
 			$deltadateforuser   = round($deltadateforclient-$deltadateforserver);
-
-			// Show logs
-			echo '<table class="border" width="100%">';
+			$userstatic = new User($log->db);
 
 			foreach ($log->lines as $line)
 			{
+				echo '<tr class="oddeven">';
+
 				// Action
-				echo '<tr><td class="titlefield">';
-				echo $langs->trans("LogAction");
-				echo '</td><td>';
-				echo $langs->trans($line->action);
-				echo '</td></tr>';
+				echo '<td align="left">'.$langs->trans($line->action).'</td>';
 
 				// Date
+				echo '<td align="center">';
 				$datec = $log->db->jdate($line->datec);
-				echo '<tr><td class="titlefield">';
-				echo $langs->trans("LogDate");
-				echo '</td><td>';
 				echo dol_print_date($datec, 'dayhour');
 				if ($deltadateforuser) {
 					echo ' '.$langs->trans("CurrentHour").' &nbsp; / &nbsp; '.dol_print_date($datec+($deltadateforuser*3600),"dayhour").' &nbsp;'.$langs->trans("ClientHour");
 				}
-				echo '</td></tr>';
+				echo '</td>';
 
 				// User
-				$userstatic = new User($log->db);
+				echo '<td align="left" colspan="2">';
 				$userstatic->fetch($line->fk_user);
-				echo '<tr><td class="titlefield">';
-				echo $langs->trans("MadeBy");
-				echo '</td><td>';
 				echo $userstatic->getNomUrl(1, '', 0, 0, 0);
-				echo '</td></tr>';
-			}
+				echo '</td>';
 
-			echo '</table>';
+				echo "</tr>\n";
+			}
 		}
 		else
 		{
-			echo '<br>'.$langs->trans('NoLogsAvailable');
+			echo '<tr><td colspan="4">'.$langs->trans('NoLogsAvailable').'</td></tr>';
 		}
+
+		// Close list
+		echo '</table>';
+		echo '</div>';
+
+		echo '</form>';
 	}
 }